--- a/csrc/quantization/cutlass_w8a8/scaled_mm_c3x.cu
+++ b/csrc/quantization/cutlass_w8a8/scaled_mm_c3x.cu
@@ -233,39 +233,43 @@ void cutlass_gemm_caller(torch::Tensor& out, torch::Tensor const& a,
   CUTLASS_CHECK(status);
 }
 
-template <typename InType, typename OutType,
-          template <typename, typename, typename> typename Epilogue, int32_t M>
-struct sm90_fp8_config {
+template <typename InType,
+          typename OutType,
+          template <typename, typename, typename> typename Epilogue>
+struct sm90_fp8_config_default {
+  // M in (128, inf) 
   static_assert(std::is_same<InType, cutlass::float_e4m3_t>());
   using KernelSchedule =
       cutlass::gemm::KernelTmaWarpSpecializedPingpongFP8FastAccum;
   using EpilogueSchedule = typename cutlass::epilogue::TmaWarpSpecialized;
   using TileShape = Shape<_128, _128, _128>;
   using ClusterShape = Shape<_2, _1, _1>;
-
   using Cutlass3xGemm =
       cutlass_3x_gemm<InType, OutType, Epilogue, TileShape, ClusterShape,
                       KernelSchedule, EpilogueSchedule>;
 };
 
-template <typename InType, typename OutType,
+template <typename InType,
+          typename OutType,
           template <typename, typename, typename> typename Epilogue>
-struct sm90_fp8_config<InType, OutType, Epilogue, 128> {
+struct sm90_fp8_config_M128 {
+  // M in (64, 128]
   static_assert(std::is_same<InType, cutlass::float_e4m3_t>());
   using KernelSchedule =
       cutlass::gemm::KernelTmaWarpSpecializedPingpongFP8FastAccum;
   using EpilogueSchedule = typename cutlass::epilogue::TmaWarpSpecialized;
   using TileShape = Shape<_64, _128, _128>;
   using ClusterShape = Shape<_2, _1, _1>;
-
   using Cutlass3xGemm =
       cutlass_3x_gemm<InType, OutType, Epilogue, TileShape, ClusterShape,
                       KernelSchedule, EpilogueSchedule>;
 };
 
-template <typename InType, typename OutType,
+template <typename InType,
+          typename OutType,
           template <typename, typename, typename> typename Epilogue>
-struct sm90_fp8_config<InType, OutType, Epilogue, 64> {
+struct sm90_fp8_config_M64 {
+  // M in [1, 64]
   static_assert(std::is_same<InType, cutlass::float_e4m3_t>());
   using KernelSchedule =
       cutlass::gemm::KernelTmaWarpSpecializedPingpongFP8FastAccum;
@@ -278,10 +282,11 @@ struct sm90_fp8_config<InType, OutType, Epilogue, 64> {
                       KernelSchedule, EpilogueSchedule>;
 };
 
-template <typename InType, typename OutType,
-          template <typename, typename, typename> typename Epilogue, int32_t M,
-          bool IsSmallN>  // IsSmallN is true if N < 8192
-struct sm90_int8_config {
+template <typename InType,
+          typename OutType,
+          template <typename, typename, typename> typename Epilogue>
+struct sm90_int8_config_default {
+  // For M > 128 and any N
   static_assert(std::is_same<InType, int8_t>());
   using KernelSchedule =
       typename cutlass::gemm::KernelTmaWarpSpecializedPingpong;
@@ -293,11 +298,11 @@ struct sm90_int8_config {
                       KernelSchedule, EpilogueSchedule>;
 };
 
-template <typename InType, typename OutType,
-          template <typename, typename, typename> typename Epilogue,
-          bool IsSmallN>
-struct sm90_int8_config<InType, OutType, Epilogue, 128, IsSmallN> {
-  // Specialization for M in (64, 128] and any N
+template <typename InType,
+          typename OutType,
+          template <typename, typename, typename> typename Epilogue>
+struct sm90_int8_config_M128 {
+  // For M in (64, 128] and any N
   static_assert(std::is_same<InType, int8_t>());
   using KernelSchedule =
       typename cutlass::gemm::KernelTmaWarpSpecializedPingpong;
@@ -309,11 +314,11 @@ struct sm90_int8_config<InType, OutType, Epilogue, 128, IsSmallN> {
                       KernelSchedule, EpilogueSchedule>;
 };
 
-template <typename InType, typename OutType,
-          template <typename, typename, typename> typename Epilogue,
-          bool IsSmallN>
-struct sm90_int8_config<InType, OutType, Epilogue, 64, IsSmallN> {
-  // Specialization for M in (32, 64] and any N
+template <typename InType,
+          typename OutType,
+          template <typename, typename, typename> typename Epilogue>
+struct sm90_int8_config_M64 {
+  // For M in (32, 64] and any N
   static_assert(std::is_same<InType, int8_t>());
   using KernelSchedule = typename cutlass::gemm::KernelTmaWarpSpecialized;
   using EpilogueSchedule = typename cutlass::epilogue::TmaWarpSpecialized;
@@ -324,10 +329,11 @@ struct sm90_int8_config<InType, OutType, Epilogue, 64, IsSmallN> {
                       KernelSchedule, EpilogueSchedule>;
 };
 
-template <typename InType, typename OutType,
+template <typename InType,
+          typename OutType,
           template <typename, typename, typename> typename Epilogue>
-struct sm90_int8_config<InType, OutType, Epilogue, 32, false> {
-  // Specialization for M in [1, 32] and N >= 8192
+struct sm90_int8_config_M32_NBig {
+  // For M in [1, 32] and N >= 8192
   static_assert(std::is_same<InType, int8_t>());
   using KernelSchedule = typename cutlass::gemm::KernelTmaWarpSpecialized;
   using EpilogueSchedule = typename cutlass::epilogue::TmaWarpSpecialized;
@@ -338,10 +344,11 @@ struct sm90_int8_config<InType, OutType, Epilogue, 32, false> {
                       KernelSchedule, EpilogueSchedule>;
 };
 
-template <typename InType, typename OutType,
+template <typename InType,
+          typename OutType,
           template <typename, typename, typename> typename Epilogue>
-struct sm90_int8_config<InType, OutType, Epilogue, 32, true> {
-  // Specialization for M in [1, 32] and N < 8192
+struct sm90_int8_config_M32_NSmall {
+  // For M in [1, 32] and N < 8192
   static_assert(std::is_same<InType, int8_t>());
   using KernelSchedule = typename cutlass::gemm::KernelTmaWarpSpecialized;
   using EpilogueSchedule = typename cutlass::epilogue::TmaWarpSpecialized;
@@ -354,24 +361,24 @@ struct sm90_int8_config<InType, OutType, Epilogue, 32, true> {
 
 }  // namespace
 
-template <typename InType, typename OutType,
+template <typename InType,
+          typename OutType,
           template <typename, typename, typename> typename Epilogue,
           typename... EpilogueArgs>
-void cutlass_gemm_sm90_fp8_dispatch(torch::Tensor& out, torch::Tensor const& a,
+void cutlass_gemm_sm90_fp8_dispatch(torch::Tensor& out,
+                                    torch::Tensor const& a,
                                     torch::Tensor const& b,
                                     EpilogueArgs&&... args) {
   static_assert(std::is_same<InType, cutlass::float_e4m3_t>());
   TORCH_CHECK(a.dtype() == torch::kFloat8_e4m3fn);
   TORCH_CHECK(b.dtype() == torch::kFloat8_e4m3fn);
 
-  static const int32_t MDimDontCare = 0;
   using Cutlass3xGemmDefault =
-      typename sm90_fp8_config<InType, OutType, Epilogue,
-                               MDimDontCare>::Cutlass3xGemm;
+      typename sm90_fp8_config_default<InType, OutType, Epilogue>::Cutlass3xGemm;
   using Cutlass3xGemmM64 =
-      typename sm90_fp8_config<InType, OutType, Epilogue, 64>::Cutlass3xGemm;
+      typename sm90_fp8_config_M64<InType, OutType, Epilogue>::Cutlass3xGemm;
   using Cutlass3xGemmM128 =
-      typename sm90_fp8_config<InType, OutType, Epilogue, 128>::Cutlass3xGemm;
+      typename sm90_fp8_config_M128<InType, OutType, Epilogue>::Cutlass3xGemm;
 
   uint32_t const m = a.size(0);
   uint32_t const mp2 =
@@ -392,38 +399,28 @@ void cutlass_gemm_sm90_fp8_dispatch(torch::Tensor& out, torch::Tensor const& a,
   }
 }
 
-template <typename InType, typename OutType,
+template <typename InType,
+          typename OutType,
           template <typename, typename, typename> typename Epilogue,
           typename... EpilogueArgs>
-void cutlass_gemm_sm90_int8_dispatch(torch::Tensor& out, torch::Tensor const& a,
+void cutlass_gemm_sm90_int8_dispatch(torch::Tensor& out,
+                                     torch::Tensor const& a,
                                      torch::Tensor const& b,
                                      EpilogueArgs&&... args) {
   static_assert(std::is_same<InType, int8_t>());
   TORCH_CHECK(a.dtype() == torch::kInt8);
   TORCH_CHECK(b.dtype() == torch::kInt8);
 
-  static const int32_t MDimDontCare = 0;
-  static const bool NDimDontCare = false;
-
-  // Same config for Large N and Small N
   using Cutlass3xGemmDefault =
-      typename sm90_int8_config<InType, OutType, Epilogue, MDimDontCare,
-                                NDimDontCare>::Cutlass3xGemm;
-  // Same config for Large N and Small N
+      typename sm90_int8_config_default<InType, OutType, Epilogue>::Cutlass3xGemm;
   using Cutlass3xGemmM128 =
-      typename sm90_int8_config<InType, OutType, Epilogue, 128,
-                                NDimDontCare>::Cutlass3xGemm;
-  // Same config for Large N and Small N
+      typename sm90_int8_config_M128<InType, OutType, Epilogue>::Cutlass3xGemm;
   using Cutlass3xGemmM64 =
-      typename sm90_int8_config<InType, OutType, Epilogue, 64,
-                                NDimDontCare>::Cutlass3xGemm;
-  // Different configs for Large N and Small N
-  using Cutlass3xGemmM32LargeN =
-      typename sm90_int8_config<InType, OutType, Epilogue, 32,
-                                false>::Cutlass3xGemm;
-  using Cutlass3xGemmM32SmallN =
-      typename sm90_int8_config<InType, OutType, Epilogue, 32,
-                                true>::Cutlass3xGemm;
+      typename sm90_int8_config_M64<InType, OutType, Epilogue>::Cutlass3xGemm;
+  using Cutlass3xGemmM32NBig =
+      typename sm90_int8_config_M32_NBig<InType, OutType, Epilogue>::Cutlass3xGemm;
+  using Cutlass3xGemmM32NSmall =
+      typename sm90_int8_config_M32_NSmall<InType, OutType, Epilogue>::Cutlass3xGemm;
 
   uint32_t const n = a.size(1);
   bool const is_small_n = n < 8192;
@@ -435,10 +432,10 @@ void cutlass_gemm_sm90_int8_dispatch(torch::Tensor& out, torch::Tensor const& a,
   if (mp2 <= 32) {
     // m in [1, 32]
     if (is_small_n) {
-      return cutlass_gemm_caller<Cutlass3xGemmM32SmallN>(
+      return cutlass_gemm_caller<Cutlass3xGemmM32NSmall>(
           out, a, b, std::forward<EpilogueArgs>(args)...);
     } else {
-      return cutlass_gemm_caller<Cutlass3xGemmM32LargeN>(
+      return cutlass_gemm_caller<Cutlass3xGemmM32NBig>(
           out, a, b, std::forward<EpilogueArgs>(args)...);
     }
   } else if (mp2 <= 64) {
@@ -456,7 +453,8 @@ void cutlass_gemm_sm90_int8_dispatch(torch::Tensor& out, torch::Tensor const& a,
   }
 }
 
-void cutlass_scaled_mm_sm90(torch::Tensor& out, torch::Tensor const& a,
+void cutlass_scaled_mm_sm90(torch::Tensor& out,
+                            torch::Tensor const& a,
                             torch::Tensor const& b,
                             torch::Tensor const& a_scales,
                             torch::Tensor const& b_scales) {