@varun-sundar-rabindranath 2024-06-13
PFA Heatmaps generated from benchmark sweeps on H100 

![model_bench-torch int8-1717210026_heatmap](https://github.com/vllm-project/vllm/assets/3337719/d343efff-e26e-4fdd-90ac-5375ec00d328)

Pointers on how to read the heatmap:
X-Axis - All the Cutlass OP tried.
Y-Axis - All the GEMM Shapes in (M x N x K) format.
The Darker the cell, the Better the algorithm performed for that GEMM shape. Each row (GEMM-Shape) is normalized to be between 0.0 and 1.0. The best algorithm has a value of 1.0.
Annotations: 
 Blue box - The configuration on `main`
 Yellow boxes - Configurations selected and added in this PR
 Green boxes - Other good configuration that were considered

Cutlass Op naming convention:
`autogen_cutlass3x_scaled_mm_dq_sm90_128x128x128_1x2x1_KernelTmaWarpSpecializedCooperative_TmaWarpSpecializedCooperative_PersistentScheduler_kGemm_int32_t_int8` refers to an Op constructed with,

TileShape : 128x128x128
ClusterShape : 1x2x1
KernelSchedule: KernelTmaWarpSpecializedCooperativeFP8FastAccum
EpilogueSchedule: TmaWarpSpecializedCooperative
TileSchedule: PersistentScheduler
Gemm mode: kGemm
int32_t_int8 : f"{accumulation_datatype}_{input_datatype}"